split backend into manageable parts
