+++ backend/controllers/healthController.js
@@ -1,0 +1,5 @@
+function getHealth(_req, res) {
+    res.json({ status: "OK" });
+}
+
+export { getHealth };
