+++ backend/package.json
@@ -10,6 +10,7 @@
     },
     "name": "streetmeetingbackend",
     "version": "1.0.0",
+    "type": "module",
     "main": "server.js",
     "scripts": {
         "start": "node server.js"
