+++ backend/routes/healthRoutes.js
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import { getHealth } from "../controllers/healthController.js";
+
+const router = Router();
+
+router.get("/health", getHealth);
+
+export { router as healthRouter };
